ajout de la fonction createThumbnail quand l'tilisateur leave est qu'il y a eu une modification dans la mindmap
